+++ src/api/RefreshAPI.ts
@@ -1,0 +1,10 @@
+import AbstractCRUDApi from "./AbstractCRUDApi";
+
+export default class RefreshAPI extends AbstractCRUDApi {
+    
+    getBaseURL(): String {
+        return "/refresh";
+     }
+}
+
+export const instance = new RefreshAPI();