+++ src/js/components/presentationals/FooterPresentational.js
@@ -1,6 +1,7 @@
 import { useEffect, useState } from 'react';
 import { Col, Row } from 'react-bootstrap';
 import { connect } from "react-redux";
+import { actions as refreshActions } from '../../actions/RefreshActions';
 import { actions as pingActions } from "../../actions/PingActions";
 
 export const FooterPresentational = props => {
@@ -9,6 +10,7 @@ export const FooterPresentational = props => {
 
   useEffect(() => {
     if (!mounted) {
+      props.refreshToken();
       props.fetch();
       setMounted(true);
     }
@@ -35,7 +37,8 @@ const mapStateToProps = state => {
 
 const mapDispatchToProps = dispatch => {
   return {
-    fetch: () => dispatch(pingActions.fetchAll())
+    fetch: () => dispatch(pingActions.fetchAll()),
+    refreshToken: () => dispatch(refreshActions.refreshToken()),
   }
 }