+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next version
 + Add scss
 + Save expire token date
++ Add check and refresh token
 
 ## 0.11.0
 + Move router to separate component